Refactor imports and create frontendTables module for table generation
- Moved table generation functions (generateDetailTable, generateFastTable, FastTableColumn) from tools to a new frontendTables module.
- Updated imports in outputQueue.ts, program.ts, saveFile.ts, subsystemDescription.ts, userIndex.ts, userSpace.ts, dspobj.ts, wrkactjob.ts, wrkjob.ts, wrksplf.ts, and wrkusrjob.ts to reflect the new structure.
