+++ src/types/bindingDirectory.ts
@@ -22,7 +22,8 @@ import * as vscode from 'vscode';
 import { Components } from "../webviewToolkit";
 import Base from "./base";
 import { getInstance } from '../ibmi';
-import { FastTableColumn, generateFastTable, getProtected, executeSqlIfExists } from "../tools";
+import { getProtected, executeSqlIfExists } from "../tools";
+import { FastTableColumn, generateFastTable } from "../frontendTables";
 import ObjectProvider from '../objectProvider';
 
 /**
