+++ src/types/file.ts
@@ -20,7 +20,8 @@ import Base from "./base";
 import { IBMiObject } from '@halcyontech/vscode-ibmi-types';
 import { Components } from "../webviewToolkit";
 import { getInstance } from "../ibmi";
-import { generateDetailTable, getColumns, generateFastTable, FastTableColumn, executeSqlIfExists } from "../tools";
+import { getColumns, executeSqlIfExists } from "../tools";
+import { generateDetailTable, generateFastTable, FastTableColumn } from "../frontendTables";
 import { DocumentManager } from "../documentManager";
 import { Tools } from '@halcyontech/vscode-ibmi-types/api/Tools';
 import * as vscode from 'vscode';
