+++ src/types/journal.ts
@@ -18,7 +18,8 @@ import Base from "./base";
 import { IBMiObject, CommandResult } from '@halcyontech/vscode-ibmi-types';
 import { Components } from "../webviewToolkit";
 import { getInstance } from "../ibmi";
-import { generateDetailTable, getColumns, generateFastTable, FastTableColumn, getProtected, checkTableFunctionExists, executeSqlIfExists } from "../tools";
+import { getColumns, getProtected, checkTableFunctionExists, executeSqlIfExists } from "../tools";
+import { generateDetailTable, generateFastTable, FastTableColumn } from "../frontendTables";
 import { DocumentManager } from "../documentManager";
 import { Tools } from '@halcyontech/vscode-ibmi-types/api/Tools';
 import * as vscode from 'vscode';
