+++ src/types/messageFile.ts
@@ -19,7 +19,8 @@ import Base from "./base";
 import { CommandResult, IBMiObject } from '@halcyontech/vscode-ibmi-types';
 import { getInstance } from "../ibmi";
 import { Tools } from '@halcyontech/vscode-ibmi-types/api/Tools';
-import { generateFastTable, FastTableColumn, executeSqlIfExists, getProtected } from "../tools";
+import { executeSqlIfExists, getProtected } from "../tools";
+import { generateFastTable, FastTableColumn } from "../frontendTables";
 import * as vscode from 'vscode';
 import ObjectProvider from "../objectProvider";
