+++ src/types/messageQueue.ts
@@ -20,7 +20,8 @@ import Base from "./base";
 import { IBMiObject, CommandResult } from '@halcyontech/vscode-ibmi-types';
 import { getInstance } from "../ibmi";
 import { Tools } from '@halcyontech/vscode-ibmi-types/api/Tools';
-import { generateFastTable, FastTableColumn, getProtected, executeSqlIfExists } from "../tools";
+import { getProtected, executeSqlIfExists } from "../tools";
+import { generateFastTable, FastTableColumn } from "../frontendTables";
 import * as vscode from 'vscode';
 import ObjectProvider from "../objectProvider";
