+++ src/types/saveFile.ts
@@ -22,7 +22,8 @@ import { Tools } from "@halcyontech/vscode-ibmi-types/api/Tools";
 import * as vscode from "vscode";
 import { getInstance } from "../ibmi";
 import ObjectProvider from '../objectProvider';
-import { executeSqlIfExists, FastTableColumn, generateDetailTable, generateFastTable, getProtected, getQSYSObjectPath } from "../tools";
+import { executeSqlIfExists, getProtected, getQSYSObjectPath } from "../tools";
+import { FastTableColumn, generateDetailTable, generateFastTable } from "../frontendTables";
 import { Components } from "../webviewToolkit";
 import Base from "./base";
 import path = require("path");
