+++ src/types/userSpace.ts
@@ -18,7 +18,8 @@
 import Base from "./base";
 import { IBMiObject } from '@halcyontech/vscode-ibmi-types';
 import { getInstance } from "../ibmi";
-import { getColumns, generateDetailTable, getProtected, executeSqlIfExists } from "../tools";
+import { getColumns, getProtected, executeSqlIfExists } from "../tools";
+import { generateDetailTable } from "../frontendTables";
 import * as vscode from 'vscode';
 import ObjectProvider from '../objectProvider';
