+++ src/views/wrkjob.ts
@@ -15,7 +15,8 @@
 
 import * as vscode from 'vscode';
 import { getInstance } from '../ibmi';
-import { FastTableColumn, generateFastTable, generateDetailTable, executeSqlIfExists } from "../tools";
+import { executeSqlIfExists } from "../tools";
+import { FastTableColumn, generateFastTable, generateDetailTable } from "../frontendTables";
 import { generatePage, Components } from "../webviewToolkit";
 import { JobOperations, SpoolOperations } from '../commonOperations';
 import { Tools } from '@halcyontech/vscode-ibmi-types/api/Tools';
